+++ eslint.config.js
@@ -11,10 +11,10 @@ export default [
 	{
 		ignores: [
 			'**/.svelte-kit',
-			'packages/adapter-static/test/apps/*/build',
-			'packages/adapter-cloudflare/files',
-			'packages/adapter-netlify/files',
-			'packages/adapter-node/files'
+			'**/test-results',
+			'**/build',
+			'**/.custom-out-dir',
+			'packages/adapter-*/files'
 		]
 	},
 	{
@@ -26,11 +26,13 @@ export default [
 		rules: {
 			'@typescript-eslint/await-thenable': 'error',
 			'@typescript-eslint/no-unused-expressions': 'off',
-			'@typescript-eslint/require-await': 'error'
+			'@typescript-eslint/require-await': 'error',
+			'@typescript-eslint/no-floating-promises': 'error'
 		},
 		ignores: [
+			'packages/adapter-cloudflare/test/apps/**/*',
 			'packages/adapter-node/rollup.config.js',
-			'packages/adapter-node/tests/smoke.spec.js',
+			'packages/adapter-node/tests/smoke.spec_disabled.js',
 			'packages/adapter-static/test/apps/**/*',
 			'packages/create-svelte/shared/**/*',
 			'packages/create-svelte/templates/**/*',