+++ packages/kit/src/runtime/server/page/index.js
@@ -1,7 +1,7 @@
 import { text } from '../../../exports/index.js';
 import { compact } from '../../../utils/array.js';
 import { get_status, normalize_error } from '../../../utils/error.js';
-import { add_data_suffix } from '../../../utils/url.js';
+import { add_data_suffix } from '../../pathname.js';
 import { Redirect } from '../../control.js';
 import { redirect_response, static_error_page, handle_error_and_jsonify } from '../utils.js';
 import {
@@ -13,10 +13,9 @@ import {
 import { load_data, load_server_data } from './load_data.js';
 import { render_response } from './render.js';
 import { respond_with_error } from './respond_with_error.js';
-import { get_option } from '../../../utils/options.js';
 import { get_data_json } from '../data/index.js';
-import { load_page_nodes } from './load_page_nodes.js';
 import { DEV } from 'esm-env';
+import { PageNodes } from '../../../utils/page_nodes.js';
 
 /**
  * The maximum request depth permitted before assuming we're stuck in an infinite loop
@@ -29,10 +28,11 @@ const MAX_DEPTH = 10;
  * @param {import('types').SSROptions} options
  * @param {import('@sveltejs/kit').SSRManifest} manifest
  * @param {import('types').SSRState} state
+ * @param {import('../../../utils/page_nodes.js').PageNodes} nodes
  * @param {import('types').RequiredResolveOptions} resolve_opts
  * @returns {Promise<Response>}
  */
-export async function render_page(event, page, options, manifest, state, resolve_opts) {
+export async function render_page(event, page, options, manifest, state, nodes, resolve_opts) {
 	if (state.depth > MAX_DEPTH) {
 		// infinite request cycle detected
 		return text(`Not found: ${event.url.pathname}`, {
@@ -46,9 +46,7 @@ export async function render_page(event, page, options, manifest, state, resolve
 	}
 
 	try {
-		const nodes = await load_page_nodes(page, manifest);
-
-		const leaf_node = /** @type {import('types').SSRNode} */ (nodes.at(-1));
+		const leaf_node = /** @type {import('types').SSRNode} */ (nodes.page());
 
 		let status = 200;
 
@@ -70,13 +68,10 @@ export async function render_page(event, page, options, manifest, state, resolve
 			}
 		}
 
-		const should_prerender_data = nodes.some((node) => node?.server?.load);
-		const data_pathname = add_data_suffix(event.url.pathname);
-
 		// it's crucial that we do this before returning the non-SSR response, otherwise
 		// SvelteKit will erroneously believe that the path has been prerendered,
 		// causing functions to be omitted from the manifest generated later
-		const should_prerender = get_option(nodes, 'prerender') ?? false;
+		const should_prerender = nodes.prerender();
 		if (should_prerender) {
 			const mod = leaf_node.server;
 			if (mod?.actions) {
@@ -93,24 +88,30 @@ export async function render_page(event, page, options, manifest, state, resolve
 		// inherit the prerender option of the page
 		state.prerender_default = should_prerender;
 
+		const should_prerender_data = nodes.should_prerender_data();
+		const data_pathname = add_data_suffix(event.url.pathname);
+
 		/** @type {import('./types.js').Fetched[]} */
 		const fetched = [];
+
+		const ssr = nodes.ssr();
+		const csr = nodes.csr();
 
 		// renders an empty 'shell' page if SSR is turned off and if there is
 		// no server data to prerender. As a result, the load functions and rendering
 		// only occur client-side.
-		if (get_option(nodes, 'ssr') === false && !(state.prerendering && should_prerender_data)) {
+		if (ssr === false && !(state.prerendering && should_prerender_data)) {
 			// if the user makes a request through a non-enhanced form, the returned value is lost
 			// because there is no SSR or client-side handling of the response
 			if (DEV && action_result && !event.request.headers.has('x-sveltekit-action')) {
 				if (action_result.type === 'error') {
 					console.warn(
-						"The form action returned an error, but +error.svelte wasn't rendered because SSR is off. To get the error page with CSR, enhance your form with `use:enhance`. See https://kit.svelte.dev/docs/form-actions#progressive-enhancement-use-enhance"
+						"The form action returned an error, but +error.svelte wasn't rendered because SSR is off. To get the error page with CSR, enhance your form with `use:enhance`. See https://svelte.dev/docs/kit/form-actions#progressive-enhancement-use-enhance"
 					);
 				} else if (action_result.data) {
 					/// case: lost data
 					console.warn(
-						"The form action returned a value, but it isn't available in `$page.form`, because SSR is off. To handle the returned value in CSR, enhance your form with `use:enhance`. See https://kit.svelte.dev/docs/form-actions#progressive-enhancement-use-enhance"
+						"The form action returned a value, but it isn't available in `page.form`, because SSR is off. To handle the returned value in CSR, enhance your form with `use:enhance`. See https://svelte.dev/docs/kit/form-actions#progressive-enhancement-use-enhance"
 					);
 				}
 			}
@@ -120,7 +121,7 @@ export async function render_page(event, page, options, manifest, state, resolve
 				fetched,
 				page_config: {
 					ssr: false,
-					csr: get_option(nodes, 'csr') ?? true
+					csr
 				},
 				status,
 				error: null,
@@ -139,7 +140,7 @@ export async function render_page(event, page, options, manifest, state, resolve
 		let load_error = null;
 
 		/** @type {Array<Promise<import('types').ServerDataNode | null>>} */
-		const server_promises = nodes.map((node, i) => {
+		const server_promises = nodes.data.map((node, i) => {
 			if (load_error) {
 				// if an error happens immediately, don't bother with the rest of the nodes
 				throw load_error;
@@ -174,10 +175,8 @@ export async function render_page(event, page, options, manifest, state, resolve
 			});
 		});
 
-		const csr = get_option(nodes, 'csr') ?? true;
-
 		/** @type {Array<Promise<Record<string, any> | null>>} */
-		const load_promises = nodes.map((node, i) => {
+		const load_promises = nodes.data.map((node, i) => {
 			if (load_error) throw load_error;
 			return Promise.resolve().then(async () => {
 				try {
@@ -208,8 +207,8 @@ export async function render_page(event, page, options, manifest, state, resolve
 		for (const p of server_promises) p.catch(() => {});
 		for (const p of load_promises) p.catch(() => {});
 
-		for (let i = 0; i < nodes.length; i += 1) {
-			const node = nodes[i];
+		for (let i = 0; i < nodes.data.length; i += 1) {
+			const node = nodes.data[i];
 
 			if (node) {
 				try {
@@ -247,16 +246,22 @@ export async function render_page(event, page, options, manifest, state, resolve
 							let j = i;
 							while (!branch[j]) j -= 1;
 
+							const layouts = compact(branch.slice(0, j + 1));
+							const nodes = new PageNodes(layouts.map((layout) => layout.node));
+
 							return await render_response({
 								event,
 								options,
 								manifest,
 								state,
 								resolve_opts,
-								page_config: { ssr: true, csr: true },
+								page_config: {
+									ssr: nodes.ssr(),
+									csr: nodes.csr()
+								},
 								status,
 								error,
-								branch: compact(branch.slice(0, j + 1)).concat({
+								branch: layouts.concat({
 									node,
 									data: null,
 									server_data: null
@@ -297,8 +302,6 @@ export async function render_page(event, page, options, manifest, state, resolve
 			});
 		}
 
-		const ssr = get_option(nodes, 'ssr') ?? true;
-
 		return await render_response({
 			event,
 			options,
@@ -306,7 +309,7 @@ export async function render_page(event, page, options, manifest, state, resolve
 			state,
 			resolve_opts,
 			page_config: {
-				csr: get_option(nodes, 'csr') ?? true,
+				csr,
 				ssr
 			},
 			status,