+++ packages/kit/src/runtime/utils.js
@@ -32,3 +32,24 @@ export function b64_encode(buffer) {
 		)
 	);
 }
+
+/**
+ * Like node's path.relative, but without using node
+ * @param {string} from
+ * @param {string} to
+ */
+export function get_relative_path(from, to) {
+	const from_parts = from.split(/[/\\]/);
+	const to_parts = to.split(/[/\\]/);
+	from_parts.pop(); // get dirname
+
+	while (from_parts[0] === to_parts[0]) {
+		from_parts.shift();
+		to_parts.shift();
+	}
+
+	let i = from_parts.length;
+	while (i--) from_parts[i] = '..';
+
+	return from_parts.concat(to_parts).join('/');
+}